requirements.txt
numpy>=1.16.2
monty
pandas
pymatgen
PyYAML 
coverage==4.3.4
codeclimate-test-reporter
fortranformat
